% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowTraceR-package.R
\docType{package}
\name{flowTraceR}
\alias{flowTraceR}
\alias{flowTraceR-package}
\title{flowTraceR: a package for standardization of level information and tracking inter-software differences in bottom-up label-free proteomics}
\description{
Useful functions to standardize software outputs from ProteomeDiscoverer, Spectronaut, DIA-NN and MaxQuant on precursor, modified peptide and proteingroup level and to trace software differences for identifications such as varying proteingroup denotations for common precursor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OKdll/flowTraceR}
}

}
\author{
\strong{Maintainer}: Oliver Kardell \email{Okdll@gmx.net}

}
